optimize module path
