+++ federatedscope/core/data/utils.py
@@ -3,11 +3,14 @@ import inspect
 import logging
 import os
 import re
-from collections import defaultdict
-from os import path as osp
+import ssl
+import urllib.request
 
 import numpy as np
+import os.path as osp
+
 from random import shuffle
+from collections import defaultdict
 
 logger = logging.getLogger(__name__)
 
@@ -648,3 +651,34 @@ def save_local_data(dir_path,
 
     if (val_data is not None) and (val_targets is not None):
         torch.save((val_data, val_targets), osp.join(dir_path, "val.pt"))
+
+
+def download_url(url: str, folder='folder'):
+    r"""Downloads the content of an url to a folder.
+
+    Modified from `https://github.com/pyg-team/pytorch_geometric/blob/master
+    /torch_geometric/data/download.py`
+
+    Args:
+        url (string): The url of target file.
+        folder (string): The target folder.
+
+    Returns:
+        path (string): File path of downloaded files.
+    """
+
+    file = url.rpartition('/')[2]
+    file = file if file[0] == '?' else file.split('?')[0]
+    path = osp.join(folder, file)
+    if osp.exists(path):
+        logger.info(f'File {file} exists, use existing file.')
+        return path
+
+    logger.info(f'Downloading {url}')
+    os.makedirs(folder, exist_ok=True)
+    ctx = ssl._create_unverified_context()
+    data = urllib.request.urlopen(url, context=ctx)
+    with open(path, 'wb') as f:
+        f.write(data.read())
+
+    return path
