+++ federatedscope/nlp/dataset/leaf_twitter.py
@@ -10,8 +10,7 @@ import os.path as osp
 from tqdm import tqdm
 from sklearn.model_selection import train_test_split
 
-from federatedscope.core.auxiliaries.utils import download_url
-from federatedscope.core.data.utils import save_local_data
+from federatedscope.core.data.utils import save_local_data, download_url
 from federatedscope.cv.dataset.leaf import LEAF, LocalDataset
 from federatedscope.nlp.dataset.utils import *
